+++ src/WebJobs.Script/Http/RetryProxyHandler.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.Azure.WebJobs.Script.Grpc
+namespace Microsoft.Azure.WebJobs.Script.Http
 {
     internal sealed class RetryProxyHandler : DelegatingHandler
     {