+++ src/WebJobs.Script/Workers/Profiles/Conditions/EnvironmentCondition.cs
@@ -30,8 +30,15 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Profiles
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _environment = environment ?? throw new ArgumentNullException(nameof(environment));
 
-            descriptor.Properties.TryGetValue(WorkerConstants.WorkerDescriptionProfileConditionName, out _name);
-            descriptor.Properties.TryGetValue(WorkerConstants.WorkerDescriptionProfileConditionExpression, out _expression);
+            if (descriptor.Properties.TryGetValue(WorkerConstants.WorkerDescriptionProfileConditionName, out var conditionNameElement))
+            {
+                _name = conditionNameElement.GetString();
+            }
+
+            if (descriptor.Properties.TryGetValue(WorkerConstants.WorkerDescriptionProfileConditionExpression, out var conditionExpressionElement))
+            {
+                _expression = conditionExpressionElement.GetString();
+            }
 
             Validate();
         }