+++ release_notes.md
@@ -3,6 +3,9 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- Improved memory metrics reporting using CGroup data for Linux consumption (#10968)
-- Fixing GrpcWorkerChannel concurrency bug (#10998)
-- Avoid circular dependency when resolving LinuxContainerLegionMetricsPublisher. (#10991)
+- Adding activity sources for Durable and WebJobs (Kafka and RabbitMQ) (#11137)
+- Add JitTrace Files for v4.1041
+- Fix startup deadlock on transient exceptions (#11142)
+- Add warning log for end of support bundle version, any bundle version < 4 (#11075), (#11160)
+- Handles loading extensions.json with empty extensions(#11174)
+- Update HttpWorkerOptions to implement IOptionsFormatter (#11175)